model/requirements.txt
tensorflow
librosa 
soundfile
flask
google-cloud-storage
gunicorn